requirements.txt
fastapi~=0.115.13
uvicorn
sqlalchemy~=2.0.41
psycopg2-binary
python-dotenv~=1.1.0
alembic~=1.16.2
pydantic[email]~=2.11.7
email-validator
passlib~=1.7.4
python-jose
python-multipart
PyJWT~=2.10.1